Refactor folder structure for navbar components
