+++ src/logged_in/components/navbar/buttons/DashboardButton.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {ROOT_URL} from "../../../shared/costants/urls";
+import {ROOT_URL} from "../../../../shared/costants/urls";
 import DashboardIcon from "@material-ui/icons/Dashboard";
 import withRouter from "react-router-dom/es/withRouter";
 import NavBarIconButton from "./NavBarIconButton";
