+++ src/logged_in/components/navbar/buttons/LogoutButton.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {ROOT_URL} from "../../../shared/costants/urls";
+import {ROOT_URL} from "../../../../shared/costants/urls";
 import {ExitToApp} from "@material-ui/icons";
 import withRouter from "react-router-dom/es/withRouter";
 import NavBarIconButton from "./NavBarIconButton";
