+++ src/logged_in/components/navbar/buttons/SchemaPageButton.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import {COVERT_URL} from "../../../shared/costants/urls";
+import {COVERT_URL} from "../../../../shared/costants/urls";
 import withRouter from "react-router-dom/es/withRouter";
 import NavBarIconButton from "./NavBarIconButton";
 import SaveAltIcon from "@material-ui/icons/SaveAlt";
