+++ src/logged_in/components/navbar/buttons/index.js
@@ -1,0 +1,6 @@
+export {default as DashboardButton} from './DashboardButton';
+export {default as EditorButton} from './EditorButton';
+export {default as FeedbackButton} from './FeedbackButton';
+export {default as LogoutButton} from './LogoutButton';
+export {default as SchemaPageButton} from './SchemaPageButton';
+export {default as SettingsButton} from './SettingsButton';
